cs_border_segment: Don't expand the radius for very small dots.
Otherwise we'd do antialiasing between the dots and it'd really look like a
solid line.

This is important for non-hidpi screens, for outlines and such in particular. I
can fix this also in the fragment shader after applying the other clips if you
think it's better, but this felt a bit less hacky.

This fixes https://bugzilla.mozilla.org/show_bug.cgi?id=1434006.